requirements.txt
streamlit
openai-whisper
transformers
torch
librosa
pyvis
networkx
spacy
ffmpeg-python
